Create Embeddings/requirements.txt
openai >= 0.27.8
python-dotenv >= 1.0.0
tiktoken >= 0.4.0